+++ lib/src/model/json/userdata.model.dart
@@ -57,6 +57,9 @@ class UserData {
   @JsonKey(defaultValue: false)
   bool requestedAudioPermission = false;
 
+  @JsonKey(defaultValue: false)
+  bool automaticallyMarkEqualMediaFilesAsOpened = false;
+
   @JsonKey(defaultValue: true)
   bool videoStabilizationEnabled = true;